Cleanup database handling
This fixes a race condition where entries would go missing
 when a transaction happened while the batch was saving as
 the batch was cleared after saving.
This is fixed by polling a queue now.

Potentially related to #60

+++ src/main/java/com/wfector/command/CleanDatabase.java
@@ -33,7 +33,7 @@ public class CleanDatabase extends BukkitRunnable {
                 where.add("`Unread`='1'");
             }
             if (cleanBefore > 0) {
-                where.add("`Time`<" + (int) (Time.getEpochTime() - cleanBefore * 24 * 60 * 60));
+                where.add("`Time`<" + (Time.getEpochTime() - cleanBefore * 24 * 60 * 60));
             }
             if (user != null) {
                 where.add("`ShopOwnerId`='" + user.toString() + "'");
